backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
kombu==5.5.4
packaging==25.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing==3.7.4.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
